+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Fixed
+
+- Fix default dropdown selector in Navbar component
+
 ## 3.6.22 (May 18, 2021)
 
 ### Added